pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leopardi-ocr"
version = "0.1.0"
description = "Research-first OCR and document parsing stack for Markdown + LaTeX output."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8",
  "pyyaml>=6.0.2",
  "rich>=14.0.0",
  "typer>=0.16.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.12.0",
]
train = [
  "accelerate>=1.7.0",
  "datasets>=3.6.0",
  "hydra-core>=1.3.2",
  "pillow>=11.2.1",
  "pymupdf>=1.26.0",
  "safetensors>=0.5.3",
  "sentencepiece>=0.2.0",
  "torch>=2.7.0",
  "transformers>=4.52.0",
  "webdataset>=0.2.111",
]

[project.scripts]
leopardi = "leopardi.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
